reorganized css
