+++ src/components/Health/HealthDetails.js
@@ -4,7 +4,8 @@ import DatasetsHealth from './DatasetsHealth'
 import CatalogSection from '../Section/CatalogSection/CatalogSection'
 import { fetchCatalog } from '../../fetch/fetch'
 import { waitForDataAndSetState, cancelAllPromises } from '../../helpers/components'
-import { container, sectionNoPadding } from './HealthDetails.css'
+import { container, sectionNoPadding, legend } from './HealthDetails.css'
+import { success, warning, error } from './Health.css'
 
 class HealthDetails extends Component {
   constructor(props) {
@@ -32,6 +33,12 @@ class HealthDetails extends Component {
         <CatalogSection catalog={catalog} />
         <UpdateHealth catalog={catalog} />
         <DatasetsHealth catalog={catalog} />
+
+        <div className={legend}>
+          <div className={error}><i className="remove icon"></i>inférieur à 20%</div>
+          <div className={warning}><i className="icon warning"></i>entre 20% et 55%</div>
+          <div className={success}><i className="checkmark icon"></i>supérieur à 5%</div>
+        </div>
 
       </div>
     </div>
